feat: refactor common modules to Esm
